backend/requirements.txt
# TrendPulse - Backend Dependencies
newsapi-python==0.2.6
requests==2.31.0
beautifulsoup4==4.12.2
tensorflow==2.14.0
transformers==4.33.2
pandas==2.0.3
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.1
python-dotenv==1.0.0
spacy==3.6.1
geopy==2.4.0
schedule==1.2.0
pydantic==2.4.2
feedparser==6.0.10
python-multipart==0.0.6
httpx==0.25.0
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy==1.11.2
joblib==1.3.2 